feat(2025d3p1): started on alternative solution
Alternative solution is to scan each line once to find the index of the largest left-most and non-final digit.
Then getting a substring of the line [idx, len) and finding the largest digit in that substring.
These digits concatenated together should be the biggest 2 digit number that can be made from the line